requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
avro==1.11.1
    # via -r requirements.in
grpcio==1.54.0
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.54.0
    # via -r requirements.in
msgpack==1.0.5
    # via -r requirements.in
protobuf==4.22.3
    # via
    #   -r requirements.in
    #   grpcio-tools
pyyaml==6.0
    # via -r requirements.in
xmltodict==0.13.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
